Started moving stuff into "node session"
